// packages/server/src/utils/types.ts
export enum Env {
	Dev = 'dev',
	Prod = 'prod',
	BuildTypes = 'buildTypes',
}

export interface NotificationView {
	id: Uuid;
	messageHtml: string;
	levelClassName: string;
	closeUrl: string;
}

interface AppContextXilinota {
	env: Env;
	db: DbConnection;
	models: Models;
	appLogger(): LoggerWrapper;
	notifications: NotificationView[];
	owner: User;
	account: Account;
	routes: Routers;
	services: Services;
}

export interface AppContext extends Koa.Context {
	xilinota: AppContextXilinota;

	// All the properties under `xilinota` were previously at the root, so to make
	// sure they are no longer used anywhere we set them to "never", as that
	// would trigger the TypeScript compiler. Later on, all this can be removed.
	// env: never;
	// db: never;
	// models: never;
	// appLogger: never;
	// notifications: never;
	// owner: never;
	// routes: never;
	// services: never;
}

export enum DatabaseConfigClient {
	Null = 'null',
	PostgreSQL = 'pg',
	SQLite = 'sqlite3',
}

export interface DatabaseConfig {
	client: DatabaseConfigClient;
	// For Postgres, this is the actual database name. For SQLite, this is the
	// path to the SQLite file.
	name: string;
	host?: string;
	port?: number;
	user?: string;
	password?: string;
	connectionString?: string;
	asyncStackTraces?: boolean;
	slowQueryLogEnabled?: boolean;
	slowQueryLogMinDuration?: number;
	autoMigration?: boolean;
}

export interface MailerConfig {
	enabled: boolean;
	host: string;
	port: number;
	security: MailerSecurity;
	authUser: string;
	authPassword: string;
	noReplyName: string;
	noReplyEmail: string;
}

export interface StripeConfig extends StripePublicConfig {
	enabled: boolean;
	secretKey: string;
	webhookSecret: string;
}

export enum StorageDriverType {
	Database = 1,
	Filesystem = 2,
	Memory = 3,
	S3 = 4,
}
// ...
export enum StorageDriverMode {
	ReadAndWrite = 1,
	ReadAndClear = 2,
}

export interface StorageDriverConfig {
	type?: StorageDriverType;
	path?: string;
	mode?: StorageDriverMode;
	region?: string;
	accessKeyId?: string;
	secretAccessKeyId?: string;
	bucket?: string;
}

export interface Config extends EnvVariables {
	appVersion: string;
	xilinotaServerVersion: string; // May be different from appVersion, if this is a fork of JS
	appName: string;
	env: Env;
	port: number;
	rootDir: string;
	viewDir: string;
	layoutDir: string;
	// Note that, for now, nothing is being logged to file. Log is just printed
	// to stdout, which is then handled by Docker own log mechanism
	logDir: string;
	tempDir: string;
	baseUrl: string;
	apiBaseUrl: string;
	adminBaseUrl: string;
	userContentBaseUrl: string;
	xilinotaAppBaseUrl: string;
	signupEnabled: boolean;
	termsEnabled: boolean;
	accountTypesEnabled: boolean;
	showErrorStackTraces: boolean;
	database: DatabaseConfig;
	mailer: MailerConfig;
	stripe: StripeConfig;
	supportEmail: string;
	supportName: string;
	businessEmail: string;
	isJoplinCloud: boolean;
	cookieSecure: boolean;
	storageDriver: StorageDriverConfig;
	storageDriverFallback: StorageDriverConfig;
	itemSizeHardLimit: number;
	maxTimeDrift: number;
}